refactor: getAttendance 개선
