+++ src/components/attendance/incompleted/InCompletedSection.tsx
@@ -1,46 +1,20 @@
 import styled from 'styled-components';
 import { Primary } from '@utils/constant/color';
 import AttendanceBox from './component/AttendanceBox';
-import { useQuery } from 'react-query';
-import { TodayAttendanceData } from '@@types/request';
-import { getAttendance } from 'src/apis/attendance';
 import Loading from '@common/loading/Loading';
-import { useRecoilValue } from 'recoil';
-import { token } from '@utils/state';
-import { useRouter } from 'next/router';
-import { AxiosError } from 'axios';
+import useAttendance from 'src/apis/queries/useAttendnace';
 
 const IncompletedSection = () => {
-  const tokens = useRecoilValue(token);
-  const router = useRouter();
-  const { data, isLoading } = useQuery<TodayAttendanceData, AxiosError>(['attendance'], () => getAttendance(tokens), {
-    retry: false,
-    onError: (error) => {
-      if (error.response?.status == 400) {
-        alert('출석 가능일이 아닙니다!');
-        router.push('/');
-      }
-      if (error.response?.status == 405) {
-        router.push('/attendance/completed');
-      }
-      if (error.response?.status == 406) {
-        alert('현재 활동 중인 아기사자가 아닙니다!');
-        router.push('/');
-      }
-    },
-    onSuccess: (data) => {
-      if (data && (data.attendance_result === 2 || data.attendance_result === 1)) {
-        router.push('/attendance/completed');
-      }
-    },
-  });
-  if (isLoading) return <Loading />;
+  const { attendance, isLoading } = useAttendance();
+
+  if (isLoading || !attendance) return <Loading />;
+
   return (
     <CircleWrapper>
       <Circle />
       <Circle2 />
       <Circle />
-      <AttendanceBox data={data!} />
+      <AttendanceBox data={attendance} />
     </CircleWrapper>
   );
 };
